+++ annotations/src/main/java/react4j/annotations/Prop.java
@@ -32,4 +32,17 @@ public @interface Prop
    */
   @Nonnull
   String name() default "<default>";
+
+  /**
+   * Return enum indicating whether prop should be when component is constructed.
+   * This influences validation when enabled and how the Builder class is created.
+   * If set to {@link Feature#ENABLE} then the user MUST supply the prop and the builder will require the user
+   * to specify the value. If set to {@link Feature#DISABLE} then the user can optionally supply the prop.
+   * If set to {@link Feature#AUTODETECT} then the annotation processor will treat it as {@link Feature#DISABLE}
+   * if theere is a corresponding {@link PropDefault} for the prop, otherwise it will be treated as
+   * {@link Feature#ENABLE}.
+   *
+   * @return the flag indicating whether the prop needs to be supplied.
+   */
+  Feature require() default Feature.AUTODETECT;
 }